requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
langchain_google_genai
langchain-community
sentence-transformers
langchain-groq 